Empty apk cache don't delete folder
When debuging it can be useful to log in to a running container and install
additional application. Without this directory you can't download the package
list though and you have to recreate it first. Since it doesn't cost anything
to keep the directory we can just remove all the files in it.